# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: AKJ Travels API
  version: 1.1.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/places:
    get:
      summary: Buscar alojamientos (paginado)
      tags: [Places]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - in: query
          name: city
          required: false
          schema: { type: string, maxLength: 30 }
        - in: query
          name: checkIn
          required: false
          schema: { type: string, format: date-time }
        - in: query
          name: checkOut
          required: false
          schema: { type: string, format: date-time }
        - in: query
          name: guestNumber
          required: false
          schema: { type: integer, minimum: 1 }
        - in: query
          name: minimum
          required: false
          schema: { type: number, format: double }
        - in: query
          name: maximum
          required: false
          schema: { type: number, format: double }
        - in: query
          name: amenities
          required: false
          description: Lista separada por comas (enum Services)
          schema: { type: string }
      responses:
        '200':
          description: Página de resultados
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponsePlacePage' }

    post:
      summary: Crear alojamiento
      tags: [Places]
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePlaceDTO' }
      responses:
        '201':
          description: Lugar creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }
  /api/places/{id}:
    put:
      summary: Editar alojamiento
      tags: [Places]
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/PathId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EditPlaceDTO' }
      responses:
        '200':
          description: Lugar editado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }

    delete:
      summary: Eliminar alojamiento (lógico)
      tags: [Places]
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/PathId' }
      responses:
        '200':
          description: Lugar eliminado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }
  /api/places/{id}/amenities:
    get:
      summary: Listar amenidades del alojamiento
      tags: [Places]
      parameters:
        - { $ref: '#/components/parameters/PathId' }
      responses:
        '200':
          description: Lista de amenidades
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseServicesList' }
  /api/places/{id}/comments:
    get:
      summary: Listar comentarios del alojamiento (recientes primero, paginado)
      tags: [Places, Comments]
      parameters:
        - { $ref: '#/components/parameters/PathId' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
      responses:
        '200':
          description: Comentarios
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseCommentPage' }
  /api/places/{bookingId}/comments:
    post:
      summary: Crear comentario (reserva COMPLETED)
      tags: [Places, Comments]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: bookingId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCommentDTO' }
      responses:
        '201':
          description: Comentario creado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }
  /api/places/{id}/bookings:
    get:
      summary: Listar reservas de un alojamiento (paginado)
      tags: [Places, Bookings]
      security: [ { bearerAuth: [] } ]
      parameters:
        - { $ref: '#/components/parameters/PathId' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - in: query
          name: state
          required: false
          schema: { $ref: '#/components/schemas/BookingState' }
      responses:
        '200':
          description: Reservas del lugar
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseBookingPage' }
  /api/places/{placeId}/stats:
    get:
      summary: Métricas del alojamiento
      tags: [Places]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: placeId
          required: true
          schema: { type: string }
        - in: query
          name: from
          required: false
          schema: { type: string, format: date-time }
        - in: query
          name: to
          required: false
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Estadísticas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponsePlaceStatsDTO' }
  /api/places/{id}/detail:
    get:
      summary: Obtener detalle de un alojamiento
      tags: [Places]
      parameters:
        - { $ref: '#/components/parameters/PathId' }
      responses:
        '200':
          description: Detalle del lugar
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponsePlaceDetailDTO' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/places/{placeId}/images:
    post:
      summary: Subir imágenes (1 a 10) y marcar principal
      tags: [Places, Images]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: placeId
          required: true
          schema: { type: string }
        - in: query
          name: mainIndex
          required: true
          schema: { type: integer, minimum: 0 }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
              required: [files]
      responses:
        '200':
          description: URLs guardadas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }

  ############################
  # BOOKINGS
  ############################
  /api/places/{placeId}/bookings:
    post:
      summary: Crear reserva para el lugar {placeId}
      tags: [Bookings]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: placeId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBookingDTO' }
      responses:
        '201':
          description: Reserva creada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResponseString' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      in: path
      name: id
      required: true
      schema: { type: string }
    Page:
      in: query
      name: page
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      in: query
      name: size
      required: false
      schema: { type: integer, minimum: 1, default: 10 }
  responses:
...
    NotFound:
      description: No encontrado
  schemas:
...
    ResponseString:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:  { type: string, example: "ok" }
...
    ResponsePlacePage:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:
          type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/PlaceEntitySummary' }
            totalElements: { type: integer, format: int64, example: 23 }
            totalPages:    { type: integer, example: 3 }
            size:          { type: integer, example: 10 }
            number:        { type: integer, example: 0 }
    ResponseBookingPage:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:
          type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/BookingDTO' }
            totalElements: { type: integer, format: int64, example: 23 }
            totalPages:    { type: integer, example: 3 }
            size:          { type: integer, example: 10 }
            number:        { type: integer, example: 0 }
    ResponsePlaceDetailDTO:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:  { $ref: '#/components/schemas/PlaceDetailDTO' }
    ResponsePlaceStatsDTO:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:  { $ref: '#/components/schemas/PlaceStatsDTO' }
    ResponseServicesList:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:
          type: array
          items: { $ref: '#/components/schemas/Services' }
    ResponseCommentPage:
      type: object
      properties:
        error: { type: boolean, example: false }
        data:
          type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/CommentDTO' }
            totalElements: { type: integer, format: int64, example: 10 }
            totalPages:    { type: integer, example: 1 }
            size:          { type: integer, example: 10 }
            number:        { type: integer, example: 0 }
...
    UserDTO:
      type: object
      properties:
        name:      { type: string }
        email:     { type: string, format: email }
        photoUrl:  { type: string, nullable: true }
        birthDate: { type: string, format: date }
        role:      { $ref: '#/components/schemas/Role' }
        createdAt: { type: string, format: date-time }
    UserDetailDTO:
      type: object
      properties:
        id:        { type: string }
        name:      { type: string }
        photoUrl:  { type: string }
        createdAt: { type: string, format: date-time }
...
    CreatePlaceDTO:
      type: object
      required: [title, description, price, placeType, capacity, country, department, city, postalCode, amenities, latitude, longitude]
      properties:
        title:       { type: string, minLength: 5, maxLength: 25 }
        description: { type: string, minLength: 20, maxLength: 500 }
        price:       { type: number, format: double, minimum: 0 }
        placeType:   { $ref: '#/components/schemas/PlaceType' }
        capacity:    { type: integer, minimum: 1, maximum: 60 }
        country:     { type: string, maxLength: 30 }
        department:  { type: string, maxLength: 30 }
        city:        { type: string, maxLength: 30 }
        neighborhood: { type: string, maxLength: 20, nullable: true }
        street:      { type: string, nullable: true }
        postalCode:  { type: string, pattern: '^[0-9A-Za-z]{4,10}$' }
        amenities:
          type: array
          items: { $ref: '#/components/schemas/Services' }
        latitude:    { type: number, format: double }
        longitude:   { type: number, format: double }
    EditPlaceDTO:
      allOf:
        - $ref: '#/components/schemas/CreatePlaceDTO'
...
    PlaceDetailDTO:
      type: object
      properties:
        id:            { type: string }
        latitude:      { type: number, format: double }
        longitude:     { type: number, format: double }
        price:         { type: number, format: double }
        picsUrl:
          type: array
          items: { type: string }
        description:   { type: string }
        amenities:
          type: array
          items: { $ref: '#/components/schemas/Services' }
        title:         { type: string }
        capacity:      { type: integer }
        averageRating: { type: number, format: double }
        user:          { $ref: '#/components/schemas/UserDetailDTO' }
    PlaceStatsDTO:
      type: object
      properties:
        reservations:  { type: integer }
        averageRating: { type: number, format: double }
    PlaceEntitySummary:
      type: object
      description: Resumen de entidad Place para favoritos/listas
      properties:
        id:            { type: string }
        title:         { type: string }
        price:         { type: number, format: double }
        city:          { type: string }
        averageRating: { type: number, format: double }
        photoUrl:      { type: string, nullable: true }

    # DTOs BOOKINGS/COMMENTS
    CreateBookingDTO:
      type: object
      required: [checkIn, checkOut, guestNumber]
      properties:
        checkIn:     { type: string, format: date-time }
        checkOut:    { type: string, format: date-time }
        guestNumber: { type: integer, minimum: 1 }
    BookingDTO:
      type: object
      properties:
        bookingState: { $ref: '#/components/schemas/BookingState' }
        user:         { $ref: '#/components/schemas/UserDTO' }
        checkIn:      { type: string, format: date-time }
        checkOut:     { type: string, format: date-time }
        guestNumber:  { type: integer }
    CreateCommentDTO:
      type: object
      required: [rating]
      properties:
        comment: { type: string, maxLength: 1000, nullable: true }
        rating:  { type: integer, minimum: 1, maximum: 5 }
    CommentDTO:
      type: object
      properties:
        user:     { $ref: '#/components/schemas/UserDTO' }
        rating:   { type: integer, minimum: 1, maximum: 5 }
        comment:  { type: string }
...
    Role:
      type: string
      enum: [USER, HOST, ADMIN]
    PlaceType:
      type: string
      enum: [HOUSE, FARM, APARTMENT]
    Services:
      type: string
      description: Enumeración extensa definida en tu proyecto
      enum:
        - BATHROOM
        - HAIR_DRYER
        - CLEANING_PRODUCTS
        - HOT_WATER
        - BEDROOM_AND_LAUNDRY
        - WASHER_IN_BUILDING_PAID
        - DRYER_IN_BUILDING_PAID
        - ESSENTIALS
        - TOWELS_BEDSHEETS_SOAP_TOILET_PAPER
        - CLOTHES_HANGERS
        - SHEETS
        - COTTON_BEDDING_MEMBER_SELECTION
        - BLACKOUT_CURTAINS
        - PARKING_AND_FACILITIES
        - FREE_STREET_PARKING
        - PAID_OFFSITE_PARKING
        - PETS_ALLOWED
        - SERVICE_ANIMALS_ALLOWED
        - LUGGAGE_DROP_OFF_ALLOWED
        - LONG_TERM_STAYS_ALLOWED
        - SELF_CHECK_IN
        - SMART_LOCK
        - NOT_INCLUDED
        - NOT_AVAILABLE_SECURITY_CAMERAS
        - NOT_AVAILABLE_AIR_CONDITIONING
        - NOT_AVAILABLE_HEATING
    BookingState:
      type: string
      enum: [PENDING, CONFIRMED, CANCELED, COMPLETED]